Add files via upload
This is an analysis report made from a data set of 491775 observations and 330 variables. I was asked to come up with 3 questions that i felt were of interest to me and come up with an analysis for to see if i could find any association between the variables selected. 

The data has been collected through landline telephonic interviews as well as cellular telephonic interviews. Their main objective is to collect uniform data that is state specific and they take into account behavioural risk factors of non-institutionalized adult population.It collects the data from all 50 states and territories.
As the interviews are conducted on a randomly selected member of a family this shows that they use stratified sampling so this means that they can only make associations from the data collected and not imply causation as all the data that is collected is observational and not obtained from a study.